requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
vosk
pyttsx3
groq
httpx
pydantic
